aruha-729: improved layering
